Add about page in project level templates folder.
I am not sure wether it is better to put all the templates in app level
template folders or in one big project level template folder but here I
am testing it out. In order to make it work you need to add
demoproject/templates to the DIRS in TEMPLATES in settings.py. For now
the about page is static but in my next commit I will add some dynamic
content to it.